+++ src/main/java/nextstep/security/web/annotation/HttpSecurityConfiguration.java
@@ -2,7 +2,6 @@ package nextstep.security.web.annotation;
 
 import nextstep.app.AuthenticationConfiguration;
 import nextstep.security.authentication.AuthenticationManager;
-import nextstep.security.context.SecurityContextHolderFilter;
 import nextstep.security.web.builders.Customizer;
 import nextstep.security.web.builders.HttpSecurity;
 import org.springframework.context.annotation.Bean;
